# src/main/java/ua/ithillel/hilleltask/model/Board.java
package ua.ithillel.hilleltask.model;

public class Board {
    private Integer id;
    private String name;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}