# src/main/java/ua/ithillel/hilleltask/model/TaskList.java
package ua.ithillel.hilleltask.model;

public class TaskList {
    private Integer id;
    private Integer boardId;
    private String name;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public Integer getBoardId() {
        return boardId;
    }

    public void setBoardId(Integer boardId) {
        this.boardId = boardId;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}