# src/main/java/ua/ithillel/hilleltask/model/dto/BoardDTO.java
package ua.ithillel.hilleltask.model.dto;

public class BoardDTO {
    private Integer id;
    private String name;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}