# src/main/java/ua/ithillel/hilleltask/model/dto/TaskDTO.java
package ua.ithillel.hilleltask.model.dto;

public class TaskDTO {
    private Integer id;
    private String title;
    private String description;
    private Integer taskListId;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public Integer getTaskListId() {
        return taskListId;
    }

    public void setTaskListId(Integer taskListId) {
        this.taskListId = taskListId;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }
}